pyproject.toml
[tool.poetry]
name = "datacollaborationanalysis"
version = "0.1.0"
description = ""
authors = ["tomos7231 <tomobake0023@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.0.1"
numpy = "^1.24.3"
seaborn = "^0.12.2"
lightgbm = "^3.3.5"
xgboost = "^1.7.5"
catboost = "^1.1.1"
black = "^23.3.0"
tqdm = "^4.65.0"
surprise = "^0.1"
category-encoders = "^2.6.0"
optuna = "^3.1.1"
pathlib = "^1.0.1"
scikit-learn = "^1.2.2"
matplotlib = "^3.7.1"
PyYAML = "^6.0"
Cython = "^0.29.34"
pyfm = "^0.2.4"

[tool.poetry.dev-dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
